+++ api/controllers/hosts/edit.js
@@ -104,10 +104,13 @@ module.exports = {
       // We must clone the InitialValues object due to how Sails.js manipulates any objects passed as InitialValues.
       var criteriaB = _.cloneDeep(inputs)
 
-      // Edit it. Also update answerCalls in recipients if necessary.
+      // Edit it. Also update answerCalls and makeCalls in recipients if necessary.
       var hostRecord = await sails.models.hosts.updateOne({ ID: inputs.ID }, criteriaB)
       if (typeof inputs.answerCalls !== 'undefined') {
         await sails.models.recipients.update({ host: `computer-${sh.unique(hostRecord.host + sails.config.custom.hostSecret)}` }, { answerCalls: inputs.answerCalls && hostRecord.authorized }).fetch()
+      }
+      if (typeof inputs.makeCalls !== 'undefined') {
+        await sails.models.recipients.update({ host: `computer-${sh.unique(hostRecord.host + sails.config.custom.hostSecret)}` }, { makeCalls: inputs.makeCalls && hostRecord.authorized }).fetch()
       }
 
       // Edit the status of this host if necessary