+++ CHANGELOG.md
@@ -62,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - forced parameter to sails.helpers.playlists.start. When true, playlist will start regardless of current meta state. [Issue 95](https://github.com/Lovinity/wwsu/issues/95)
  - "actual" property to planner/add and planner/edit for setting and modifying final scheduled times for shows.
  - "answerCalls" boolean property to recipients. It is true when the corresponding host's authorized and answerCalls settings are true, otherwise it is false.
+ - "makeCalls" boolean property to recipients. It is true when the corresponding host's authorized and makeCalls settings are true, otherwise it is false.
 
 ### Changed
  - Many of the methods used in api controllers and in models have been migrated to sails helpers.